refactor: extract CallGraphService, SymbolIndexStore, and CompilationEnvironment from CppAnalyzer
Decompose CppAnalyzer monolith by extracting three cohesive components:

- CallGraphService: manages call graph analysis (incoming/outgoing calls, call sites)
- SymbolIndexStore: manages symbol indexes (class, function, file, USR) and file hashes
- CompilationEnvironment: manages compilation setup (compile commands, file scanner, compile args)

CppAnalyzer retains its public API through backward-compatible properties and delegation methods,
ensuring no changes needed to MCP server layer or tests. All 1435 tests pass.
